test(sessions): split canonical retention coverage
